+++ docs/changes.rst
@@ -26,6 +26,13 @@ v0.8 (in development)
   known setting, and suggests to the user what we think the setting should have
   been.
 
+- Removed most traces of multiple outputs support. Having this feature
+  currently seems to be more trouble than what it is worth.
+  :attr:`mopidy.settings.OUTPUTS` setting is no longer supported, and has been
+  replaced with :attr:`mopidy.settings.OUTPUT` which is a GStreamer
+  bin described in the same format as ``gst-launch`` expects. Default value is
+  ``autoaudiosink``.
+
 
 v0.7.3 (2012-08-11)
 ===================